♻️(tests) move referral unassign unit tests to a separate file
The test_api_referral.py file for all tests related to the referral
model has grown far too large and unwieldy. We're splitting it to
make those tests easier to run in isolated groups.
